Rename drq utils
