+++ torchbenchmark/models/drq/drq.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 import copy
 import math
 
-from . import utils
+from . import drqutils
 
 class Encoder(nn.Module):
     """Convolutional encoder for image-based observations."""
@@ -65,7 +65,7 @@ class Encoder(nn.Module):
     def copy_conv_weights_from(self, source):
         """Tie convolutional layers"""
         for i in range(self.num_layers):
-            utils.tie_weights(src=source.convs[i], trg=self.convs[i])
+            drqutils.tie_weights(src=source.convs[i], trg=self.convs[i])
 
     def log(self, logger, step):
         pass
@@ -79,11 +79,11 @@ class Actor(nn.Module):
         self.encoder = Encoder(*encoder_cfg)
 
         self.log_std_bounds = log_std_bounds
-        self.trunk = utils.mlp(self.encoder.feature_dim, hidden_dim,
+        self.trunk = drqutils.mlp(self.encoder.feature_dim, hidden_dim,
                                2 * action_shape[0], hidden_depth)
 
         self.outputs = dict()
-        self.apply(utils.weight_init)
+        self.apply(drqutils.weight_init)
 
     def forward(self, obs, detach_encoder=False):
         obs = self.encoder(obs, detach=detach_encoder)
@@ -100,7 +100,7 @@ class Actor(nn.Module):
         self.outputs['mu'] = mu
         self.outputs['std'] = std
 
-        dist = utils.SquashedNormal(mu, std)
+        dist = drqutils.SquashedNormal(mu, std)
         return dist
 
     def log(self, logger, step):
@@ -113,13 +113,13 @@ class Critic(nn.Module):
 
         self.encoder = Encoder(*encoder_cfg)
 
-        self.Q1 = utils.mlp(self.encoder.feature_dim + action_shape[0],
+        self.Q1 = drqutils.mlp(self.encoder.feature_dim + action_shape[0],
                             hidden_dim, 1, hidden_depth)
-        self.Q2 = utils.mlp(self.encoder.feature_dim + action_shape[0],
+        self.Q2 = drqutils.mlp(self.encoder.feature_dim + action_shape[0],
                             hidden_dim, 1, hidden_depth)
 
         self.outputs = dict()
-        self.apply(utils.weight_init)
+        self.apply(drqutils.weight_init)
 
     def forward(self, obs, action, detach_encoder=False):
         assert obs.size(0) == action.size(0)
@@ -197,7 +197,7 @@ class DRQAgent(object):
         action = dist.sample() if sample else dist.mean
         action = action.clamp(*self.action_range)
         assert action.ndim == 2 and action.shape[0] == 1
-        return utils.to_np(action[0])
+        return drqutils.to_np(action[0])
 
     def update_critic(self, obs, obs_aug, action, reward, next_obs,
                       next_obs_aug, not_done, logger, step):
@@ -277,5 +277,5 @@ class DRQAgent(object):
             self.update_actor_and_alpha(obs, logger, step)
 
         if step % self.critic_target_update_frequency == 0:
-            utils.soft_update_params(self.critic, self.critic_target,
+            drqutils.soft_update_params(self.critic, self.critic_target,
                                      self.critic_tau)
